Add custom id generator
The JpaDataGenerator can now be used to generate an object graph with
an artifical, unique id for each entity without having to perist the
entity and retrieve the id assigned from the database. This is useful
for when using the generator in conjunction with the JpaToMsConverter,
as MS has no facilities for generating artifical identifiers.
Obviously the generator should never generate artifical ids for the
objects when it is desired to write the objects to the relational
database, as this may lead to issues.