Reduce usage of eager storing by employing field evaluation
Implement an evaluator which designates the simple fields of entities
that are modified by the transactions of the business model to be
stored eagerly (while using the default lazy storing), meaning it is
not necessary to use an eager storer in all cases where an existing
object is modified. This should in some scenarios result in a
performance improvement.