Fix performance issue with MS new order service
Previously, the new order transaction would sometimes take multiple
seconds to complete. This was due to the employment of eager storing,
which resulted in massive amounts of data being written, while only a
very small amount of data was actually being modified.
Defaulting back to default, lazy storing resulted in better performance
while only storing the data that actually has to be udpated.