Migrate MS data access from custom impl to JACIS
This commit makes a huge modification to how persistence with
MicroStream is handled; instead of simply using a global lock to
prevent concurrent threads from working on the same persistent data at
the same time, this implementation uses the JACIS framework to add
basic transaction support to the persistence implementation using
MicroStream. This is achieved by modifying the data model in such a way
that it is more similar to how data is being kept in a relational
database; objects no longer store references to each other. Instead,
they only hold ids. The major object types each are managed by a
JacisStore, which enables more granular locking and a generic rollback
mechanism to be employed.

For JACIS to work, some control had to be relinquished; MicroStream
data access is no longer responsibility of any custom code; instead,
this is now handled by JACIS, which uses its own data root to map the
store objects to MicroStream storage.

The decision to employ JACIS was made as otherwise I would have had to
make the data model more immutable, which is impossible with the
current business model, as basically every object in the data model
hierarchy is modified at some point.

Note that JACIS was added as a local, in-project dependency, as the
most recent published version (2.0.6) uses a rather outdated version of
MicroStream. The local dependency is a custom compilation of the JACIS
library which uses MicroStream 5+.