Remove automatic MS storage clearing
Remove the option to enable the automatic clearing of the files created
by MicroStream at application startup. An alternative implementation
which attempted to delete the directory used for data storage right
before the storage manager started was not added, as this interestingly
never resulted in the actual removal of the files on the system, even
when using well-known implementations such as Spring's FileSystemUtils
methods. This might be due to the fact that MicroStream itself attempts
to create the directory right after the deletion and some background
magic (i.e. maybe the initial delete-order are deferred and nullified
for some reason...).