Custom JacisReadonlyModeSupport implementation
Add a custom implementation of of the JaciReadonlyModeSupport
interface, which does not rely on thread checking. Instead, this
implementation uses a simple boolean flag to determine whether an
object is writable or not. The performance of this implementation is
identical to that of the JacisStores.fastStream(store, filter) helper
method, which thuws was removed with this commit.